// src/crossdatabase.client/src/components/node/models/NodeData.ts
import { NodeType } from '../enums/nodeTypes.ts';

export type NodeData = {
    nodes: NodeElement[];
    connectors: Connector[];
};

export type NodeElement = {
    id: number;
    type: NodeType;
    name: string;
    posX: number;
    posY: number;
    fields: Fields;
};

export type Connector = {
    id: number;
    fromNode: number;
    from: string;
    toNode: number;
    to: string;
};

export type Fields = {
    inputs: string[];
    outputs: string[];
};